Renamed boore_atkinson_2008->atkinson_boore_2006
