+++ openquake/hazardlib/gsim/atkinson_boore_2006.py
