+++ openquake/hazardlib/gsim/can15/eastern.py
@@ -13,7 +13,7 @@ from openquake.hazardlib.imt import PGA, SA
 from openquake.hazardlib.const import StdDev, IMC, TRT
 from openquake.hazardlib.gsim.pezeshk_2011 import PezeshkEtAl2011
 from openquake.hazardlib.gsim.boore_atkinson_2011 import Atkinson2008prime
-from openquake.hazardlib.gsim.boore_atkinson_2008 import \
+from openquake.hazardlib.gsim.atkinson_boore_2006 import \
     AtkinsonBoore2006Modified2011
 from openquake.hazardlib.gsim.silva_2002 import (
     SilvaEtAl2002SingleCornerSaturation, SilvaEtAl2002DoubleCornerSaturation)
